pyproject.toml
[project]
name = "okx-quant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "websockets>=12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "tabulate>=0.9.0",
]
